Enhances ROS2 adapter stability and outlines RViz visualization plan
Mitigates `rmw_fastrtps` DDS graph deadlocks by moving discovery queries to a dedicated background thread. This avoids contention when retrieving ROS2 topic, service, and node information.

Centralizes and improves topic type resolution across components, with plot managers, expression plots, and service callers now leveraging the robust, non-blocking `TopicDiscovery` cache. This component now includes startup grace periods and topology change cooldowns to prevent hammering the DDS layer during active discovery.

Further refines message introspection by correcting dynamic array size calculation in the Topic Echo panel, leading to more accurate data display for complex message types.

Introduces `plans/SPECTRA_ROS_STUDIO_PLAN.md`, a comprehensive document outlining the architecture, roadmap, and design principles for evolving the application into an RViz-class 3D visualization tool. This plan details the phased implementation of new display plugins, a scene manager, and an advanced TF buffer.

+++ src/ui/imgui/imgui_integration.cpp
@@ -143,7 +143,7 @@ bool ImGuiIntegration::init(VulkanBackend& backend, GLFWwindow* window, bool ins
     ImGui_ImplVulkan_Init(&ii);
     ImGui_ImplVulkan_CreateFontsTexture();
 
-    cached_render_pass_ = reinterpret_cast<uint64_t>(ii.RenderPass);
+    cached_render_pass_ = reinterpret_cast<uintptr_t>(ii.RenderPass);
     initialized_        = true;
     return true;
 }
@@ -203,7 +203,7 @@ bool ImGuiIntegration::init_headless(VulkanBackend& backend, uint32_t width, uin
     ImGui_ImplVulkan_Init(&ii);
     ImGui_ImplVulkan_CreateFontsTexture();
 
-    cached_render_pass_ = reinterpret_cast<uint64_t>(ii.RenderPass);
+    cached_render_pass_ = reinterpret_cast<uintptr_t>(ii.RenderPass);
     initialized_        = true;
     return true;
 }
@@ -262,7 +262,7 @@ void ImGuiIntegration::on_swapchain_recreated(VulkanBackend& backend)
     // Vulkan backend to pick up the new render pass.  This is a no-op in the
     // common case where recreate_swapchain reuses the render pass handle.
     VkRenderPass current_rp      = backend.render_pass();
-    auto         current_rp_bits = reinterpret_cast<uint64_t>(current_rp);
+    auto         current_rp_bits = reinterpret_cast<uintptr_t>(current_rp);
     if (current_rp_bits != cached_render_pass_ && current_rp != VK_NULL_HANDLE)
     {
         SPECTRA_LOG_WARN(
